src/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "usepolvo"
version = "1.0.1"
description = "A comprehensive integration package supporting various services."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp==3.11.12",
    "asyncio==3.4.3",
    "cryptography==44.0.1",
    "gql==3.5.1",
    "pydantic-settings==2.7.1",
    "pydantic[email]==2.10.6",
    "python-dotenv==1.0.1",
    "ratelimit==2.2.1",
    "requests==2.32.3",
    "tenacity==9.0.0",
]
[[project.authors]]
name = "Caio Pizzol"
email = "hi@usepolvo.com"

[project.optional-dependencies]
hubspot = ["hubspot-api-client==9.0.0"]
gemini = ["google-genai==0.10.0"]
openai = ["openai==1.63.2"]
anthropic = ["anthropic==0.45.2"]
all = [
    "hubspot-api-client==9.0.0",
    "google-genai==0.10.0",
    "openai==1.63.2",
    "anthropic==0.45.2",
]

[project.license]
text = "MIT"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "--maxfail=2 -ra"
testpaths = ["tests"]
